+++ packages/entity-database-adapter-knex/src/__integration-tests__/PostgresEntityIntegration-test.ts
@@ -157,6 +157,28 @@ describe('postgres entity integration', () => {
     });
   });
 
+  describe('BIGINT fields', () => {
+    it('supports BIGINT fields', async () => {
+      const vc1 = new ViewerContext(createKnexIntegrationTestEntityCompanionProvider(knexInstance));
+
+      const entity = await enforceAsyncResult(
+        PostgresTestEntity.creator(vc1).setField('bigintField', '72057594037928038').createAsync()
+      );
+
+      expect(entity.getField('bigintField')).toEqual('72057594037928038');
+    });
+
+    it('supports small ints in BIGINT field (still as string though)', async () => {
+      const vc1 = new ViewerContext(createKnexIntegrationTestEntityCompanionProvider(knexInstance));
+
+      const entity = await enforceAsyncResult(
+        PostgresTestEntity.creator(vc1).setField('bigintField', '10').createAsync()
+      );
+
+      expect(entity.getField('bigintField')).toEqual('10');
+    });
+  });
+
   describe('conjunction field equality loading', () => {
     it('supports single fieldValue and multiple fieldValues', async () => {
       const vc1 = new ViewerContext(createKnexIntegrationTestEntityCompanionProvider(knexInstance));
